Create n queens method.c
The N Queens Problem is a classic problem in computer science that involves placing N chess queens on an NxN chessboard in such a way that no two queens threaten each other. That is, no two queens are placed in the same row, column, or diagonal.

One way to solve the N Queens Problem is to use backtracking, which involves recursively placing queens on the board and undoing the placement if a solution cannot be found. Here is an example implementation of the N Queens Problem in C using the backtracking method:

c
Copy code